+++ scala/colorful/colorful-lib/src/main/scala/x7c1/colorful/lib/chapter15/former/FileOperation.scala
@@ -1,5 +1,4 @@
-package x7c1.colorful.lib.chapter15
-
+package x7c1.colorful.lib.chapter15.former
 
 sealed trait FileOperation[A]
